+++ CHANGELOG.md
@@ -1,5 +1,5 @@
 # Verb Changelog
-## master
+## **2.8.0** - 2020-02-23
 - All the `verb-send-request-on-point-*` commands now accept a prefix argument (<kbd>C-u</kbd>). When it is specified, the user can modify the request that was extracted from the headings hierarchy before it is sent, in a temporary buffer.
 - The `verb-set-var` command can now set variable values for variables that haven't been used before.
 - Changed type of `verb-content-type-handlers` elements (to lists).